# swagger/src/swagger.json
{
	"components": {
...
		"schemas": {
...
			"CountResponse": {
				"description": "Response for general count",
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"count"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UUID": {
				"type": "string",
				"example": "52907745-7672-470e-a803-a2f8feb52944",
				"description": "Stringified UUIDv4.\nSee [RFC 4112](https://tools.ietf.org/html/rfc4122)",
				"pattern": "[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-4[0-9A-Fa-f]{3}-[89ABab][0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}"
			},
...
			"email": {
				"type": "string",
				"example": "member@vps.com",
				"pattern": "[a-z0-9]+@[a-z]+\\.[a-z]{2,3}"
			},
			"LoginResponse": {
				"description": "response of login api",
				"properties": {
					"jwt": {
						"type": "string"
					},
					"user": {
						"properties": {
							"avatar": {
								"type": "string"
							},
							"nxToken": {
								"type": "string"
							},
							"tokenExpiredAt": {
								"type": "string",
								"format": "date-time"
							},
							"latestLogin": {
								"type": "string",
								"format": "date-time"
							},
							"role": {
								"$ref": "#/components/schemas/UUID"
							},
							"email": {
								"$ref": "#/components/schemas/email"
							},
							"name": {
								"type": "string"
							},
							"id": {
								"$ref": "#/components/schemas/UUID"
							}
						},
						"required": [
							"avatar",
							"nxToken",
							"tokenExpiredAt",
							"latestLogin",
							"role",
							"email",
							"name",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"jwt",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginRequest": {
				"description": "request of login api",
				"properties": {
					"email": {
						"$ref": "#/components/schemas/email",
						"description": "email of user account"
					},
					"password": {
						"type": "string",
						"description": "password of user account"
					},
					"nxWitnessLogin": {
						"type": "boolean",
						"description": "is user using NX Witness Account?"
					}
				},
				"required": [
					"email",
					"password",
					"nxWitnessLogin"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponse": {
				"description": "success response on logout",
				"properties": {
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserResponse": {
				"description": "response of user apis",
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UUID"
					},
					"nxWitnessId": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"$ref": "#/components/schemas/email"
					},
					"callingCode": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"roleId": {
						"$ref": "#/components/schemas/UUID"
					},
					"status": {
						"type": "string"
					},
					"isVerified": {
						"type": "boolean"
					},
					"blockingReason": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"latestLogin": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"nxWitnessId",
					"avatar",
					"name",
					"email",
					"callingCode",
					"phoneNumber",
					"roleId",
					"status",
					"isVerified",
					"blockingReason",
					"role",
					"latestLogin"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserList": {
				"description": "user list as response",
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/UserResponse"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserUpdateBody": {
				"description": "request body of updating user information",
				"properties": {
					"name": {
						"type": "string"
					},
					"roleId": {
						"$ref": "#/components/schemas/UUID"
					},
					"callingCode": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"avatarUrl": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"roleId",
					"callingCode",
					"phoneNumber",
					"avatarUrl"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyEmailBody": {
				"description": "request body on verifying user's email",
				"properties": {
					"email": {
						"$ref": "#/components/schemas/email"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BooleanResponse": {
				"description": "success response on logout",
				"properties": {
					"data": {
						"type": "boolean"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserCreationBody": {
				"description": "request body of creating user",
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"$ref": "#/components/schemas/email"
					},
					"roleId": {
						"$ref": "#/components/schemas/UUID"
					},
					"callingCode": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"roleId",
					"callingCode",
					"phoneNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ForgetPasswordResponse": {
				"description": "Response when asking for forget password",
				"properties": {
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"$ref": "#/components/schemas/UserResponse"
					}
				},
				"required": [
					"id",
					"userId",
					"createdAt",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"Authorization": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "Bearer Token get from Login API (JWT)"
			}
		}
	},
	"info": {
		"title": "swagger-service",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "VPS"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/user/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					}
				}
			}
		},
		"/user/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": []
			}
		},
		"/user/list": {
			"get": {
				"operationId": "ListUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserList"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": []
			}
		},
		"/user/count": {
			"get": {
				"operationId": "CountUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CountResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": []
			}
		},
		"/user/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdateBody"
							}
						}
					}
				}
			}
		},
		"/user": {
			"put": {
				"operationId": "SelfUpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdateBody"
							}
						}
					}
				}
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreationBody"
							}
						}
					}
				}
			}
		},
		"/user/verify-email/{id}": {
			"post": {
				"operationId": "VerifyEmailForUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifyEmailBody"
							}
						}
					}
				}
			}
		},
		"/user/forget-password": {
			"post": {
				"operationId": "CreateForgetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BooleanResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreationBody"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "CheckForgetPasswordKey",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForgetPasswordResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/block": {
			"put": {
				"operationId": "BlockUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/activate": {
			"put": {
				"operationId": "ActivateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:9000/apis",
			"description": "Local environment"
		}
	]
}